Move PageType to its own class
